+++ tests/viisp/test_forms.py
@@ -6,13 +6,14 @@ from vitrina.users.factories import UserFactory
 @pytest.mark.django_db
 def test_fake_viisp_form_valid_email():
     user = UserFactory(email="valid@test.com")
-    form = FakeViispForm(data={"email": user.email})
+    user.set_password("test")
+    user.save()
+    form = FakeViispForm(data={"username": user.email, "password": "test"})
     assert form.is_valid()
 
 
 @pytest.mark.django_db
 def test_fake_viisp_form_invalid_email():
-    # Email not in DB
-    form = FakeViispForm(data={"email": "not.existing@test.com"})
+    form = FakeViispForm(data={"username": "not.existing@test.com", "password": "abc"})
     assert not form.is_valid()
-    assert "Naudotojas su tokiu el. paštu neegzistuoja." in form.errors["email"][0]
+    assert "Įveskite teisingą Elektroninis paštas ir slaptažodį. Abiejuose laukuose didžiosios mažosios raidės skiriasi." in form.errors["__all__"]