+++ vitrina/viisp/forms.py
@@ -1,10 +1,9 @@
 from django import forms
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Field, Submit
-from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
+from django.contrib.auth.forms import AuthenticationForm
 
-from vitrina.users.models import User
 
 PROXY_TYPE_CHOICES = [
     ("generic", "Generic"),
@@ -14,8 +13,7 @@ PROXY_TYPE_CHOICES = [
 ]
 
 
-class FakeViispForm(forms.Form):
-    email = forms.EmailField(label=_("El. paštas"), required=True)
+class FakeViispForm(AuthenticationForm):
     lt_company_code = forms.CharField(
         label=_("Įmonės kodas"),
         help_text=_("Įveskite Lietuvoje registruotos įmonės kodą."),
@@ -28,14 +26,9 @@ class FakeViispForm(forms.Form):
         self.helper = FormHelper()
         self.helper.form_id = "fake-viisp-form"
         self.helper.layout = Layout(
-            Field("email"),
+            Field("username"),
+            Field("password"),
             Field("lt_company_code"),
             Field("proxy_type"),
             Submit("submit", _("Prisijungti"), css_class="button is-primary"),
         )
-
-    def clean_email(self) -> str:
-        email = self.cleaned_data.get("email")
-        if not User.objects.filter(email=email).exists():
-            raise ValidationError(_("Naudotojas su tokiu el. paštu neegzistuoja."))
-        return email