+++ vitrina/viisp/views.py
@@ -1,9 +1,8 @@
-from django.http import HttpRequest, HttpResponse
+from django.http import HttpResponse
 from django.views import View
 from django.views.generic import TemplateView
 from django.urls import reverse
 from django.shortcuts import render, redirect
-from django.contrib.auth import login as django_login
 from allauth.socialaccount import providers
 from allauth.socialaccount.providers.oauth2.views import (
     OAuth2LoginView,
@@ -27,6 +26,7 @@ from allauth.account.utils import perform_login
 from cryptography.fernet import Fernet
 from allauth.socialaccount.models import SocialAccount
 from itsdangerous.url_safe import URLSafeSerializer
+from django.contrib.auth.views import LoginView
 import bcrypt
 
 
@@ -147,37 +147,24 @@ class VIISPCompleteLoginView(View):
         return response
 
 
-class FakeVIISPCompleteLoginView(View):
+class FakeVIISPCompleteLoginView(LoginView):
     """Fake VIISP login for debug/testing purposes only."""
 
+    template_name = "vitrina/viisp/fake_viisp_form.html"
     form_class = FakeViispForm
-    cleaned_data = None
+    redirect_authenticated_user = True
 
-    def get(self, request: HttpRequest) -> HttpResponse:
-        form = FakeViispForm()
-        return render(request, "vitrina/viisp/fake_viisp_form.html", {"form": form})
+    def form_valid(self, form) -> HttpResponse:
+        user: User = form.get_user()
+        user.is_viisp_login = True
+        if company_code := form.cleaned_data.get("lt_company_code"):
+            user.viisp_company_code = company_code
+        else:
+            user.viisp_company_code = None
 
-    def post(self, request: HttpRequest) -> HttpResponse:
-        form: FakeViispForm = self.form_class(request.POST)
+        user.save()
 
-        if form.is_valid():
-            email: str = form.cleaned_data["email"]
-            company_code: int = form.cleaned_data.get("lt_company_code")
-            user, _ = User.objects.get_or_create(email=email)
-            user.is_viisp_login = True
-            if company_code:
-                user.viisp_company_code = company_code
-            else:
-                user.viisp_company_code = None
-            user.save()
-            django_login(request, user, backend="django.contrib.auth.backends.ModelBackend")
-            return redirect("home")
-
-        return render(
-            request,
-            "vitrina/viisp/fake_viisp_form.html",
-            {"form": form},
-        )
+        return super().form_valid(form)
 
 
 def _confirm_viisp_email(