+++ CHANGES.rst
@@ -10,6 +10,9 @@ https://github.com/atviriduomenys/katalogas/issues/2123
 - Update partner registration permissions: validate that user is in active VIISP session.
 - Update fake VIISP login to work with new partner registration.
 
+https://github.com/atviriduomenys/dvms/issues/346
+- Add password validation for fake VIISP login.
+
 v 1.7 (2025-11-20)
 ==================